Final commit
Not happy. The scheduler works perfectly but I don't think the stat gathering is correct. And the user rolls to decide how long to be blocked on an event are always identical and I have no idea how. All my other rolls using the same function are randomizing perfectly. Here's hoping for an A-.